pyproject.toml
[tool.poetry]
name = "portafolio"
version = "0.1.0"
description = "Mi portafolio"
authors = ["Engel Pinto <engeljavierpinto@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
wagtail = "^2.12.4"
python-decouple = "^3.4"
whitenoise = "^5.2.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
